Fix various printf() issues: too many arguments, not enough arguments, argument order reversed, wrong modifiers. ok deraadt@ marco@ mickey@

+++ sys/dev/pcmcia/aic_pcmcia.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: aic_pcmcia.c,v 1.11 2005/01/27 17:04:55 millert Exp $	*/
+/*	$OpenBSD: aic_pcmcia.c,v 1.12 2005/11/21 18:16:41 millert Exp $	*/
 /*	$NetBSD: aic_pcmcia.c,v 1.6 1998/07/19 17:28:15 christos Exp $	*/
 
 /*
@@ -153,7 +153,8 @@ aic_pcmcia_attach(parent, self, aux)
 		return;
 	}
 
-	printf(" port 0x%lx/%d", psc->sc_pcioh.addr, psc->sc_pcioh.size);
+	printf(" port 0x%lx/%lu", psc->sc_pcioh.addr,
+	    (u_long)psc->sc_pcioh.size);
 
 	if (!aic_find(sc->sc_iot, sc->sc_ioh)) {
 		printf(": unable to detect chip!\n");
